aiovici/requirements.txt
asyncinit
aioserial
pyserial
